# Back-end et cloud/API/swagger/v1/spec.json
{"openapi":"3.0.0","info":{"title":"API","version":"1.0.0"},"paths":{"/sponsoring/getAllSponsoring/{order}/{iPage}":{"get":{"security":[{"bearerAuth":[]}],"tags":["Sponsoring"],"parameters":[{"in":"path","name":"order","schema":{"type":"string"},"required":true,"description":"The name of the column by which the array should be sorted"},{"in":"path","name":"iPage","schema":{"type":"integer"},"required":true,"description":"The page number"}],"responses":{"200":{"description":"Returns a JSON array of 10 sponsoring, ordered by the column name chosen.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/getAllSponsoring"}}}},"401":{"$ref":"#/components/responses/UnauthorizedError"},"403":{"$ref":"#/components/responses/mustBeAdmin"},"404":{"description":"No sponsoring.","content":{"text/plain":{"schema":{"type":"string"}}}},"500":{"description":"Error while connecting to database."}}}},"/sponsoring/getAllSponsoringAndPagesCount/{order}/{iPage}":{"get":{"security":[{"bearerAuth":[]}],"tags":["Sponsoring"],"parameters":[{"in":"path","name":"order","schema":{"type":"string"},"required":true,"description":"The name of the column by which the array should be sorted"},{"in":"path","name":"iPage","schema":{"type":"integer"},"required":true,"description":"The page number"}],"responses":{"200":{"description":"Returns a JSON object:<br> .sponsoring: a JSON array of 10 sponsoring, ordered by the column name chosen<br> .nbPagesSponsoring: The number of pages","content":{"application/json":{"schema":{"$ref":"#/components/schemas/getAllSponsoringAndPagesCount"}}}},"401":{"$ref":"#/components/responses/UnauthorizedError"},"403":{"$ref":"#/components/responses/mustBeAdmin"},"404":{"description":"No sponsoring.","content":{"text/plain":{"schema":{"type":"string"}}}},"500":{"description":"Error while connecting to database."}}}},"/sponsoring/getSearchSponsoring/{search}/{order}/{iPage}":{"get":{"security":[{"bearerAuth":[]}],"tags":["Sponsoring"],"parameters":[{"in":"path","name":"search","schema":{"type":"string"},"required":true,"description":"The value of the research"},{"in":"path","name":"order","schema":{"type":"string"},"required":true,"description":"The name of the column by which the array should be sorted"},{"in":"path","name":"iPage","schema":{"type":"integer"},"required":true,"description":"The page number"}],"responses":{"200":{"description":"Returns a JSON object result:<br> .sponsoring: a JSON array of 10 sponsoring matching the search value, ordered by the column name chosen <br> .nbPagesSponsoring: The number of pages","content":{"application/json":{"schema":{"$ref":"#/components/schemas/getAllSponsoringAndPagesCount"}}}},"401":{"$ref":"#/components/responses/UnauthorizedError"},"403":{"$ref":"#/components/responses/mustBeAdmin"},"404":{"description":"No sponsoring.","content":{"text/plain":{"schema":{"type":"string"}}}},"500":{"description":"Internal server error."}}}},"/sponsoring/add":{"post":{"security":[{"bearerAuth":[]}],"tags":["Sponsoring"],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/addSponsoring"}}}},"responses":{"201":{"$ref":"#/components/responses/sponsoringAdded"},"400":{"description":"the error(s) described","content":{"text/plain":{"schema":{"type":"string"}}}},"401":{"$ref":"#/components/responses/UnauthorizedError"},"403":{"$ref":"#/components/responses/mustBeAdmin"},"500":{"description":"internal server error."}}}},"/sponsoring/update":{"patch":{"security":[{"bearerAuth":[]}],"tags":["Sponsoring"],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/updateSponsoring"}}}},"responses":{"200":{"description":"Returns a JSON array of 10 sponsoring, ordered by the column name chosen.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/getAllSponsoring"}}}},"401":{"$ref":"#/components/responses/UnauthorizedError"},"403":{"$ref":"#/components/responses/mustBeAdmin"},"404":{"description":"No sponsoring.","content":{"text/plain":{"schema":{"type":"string"}}}},"500":{"description":"Error while connecting to database."}}}},"/sponsoring/delete":{"delete":{"security":[{"bearerAuth":[]}],"tags":["Sponsoring"],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/deleteSchema"}}}},"responses":{"200":{"description":"Returns a JSON array of 10 sponsoring, ordered by the column name chosen.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/getAllSponsoring"}}}},"401":{"$ref":"#/components/responses/UnauthorizedError"},"403":{"$ref":"#/components/responses/mustBeAdmin"},"404":{"description":"No sponsoring.","content":{"text/plain":{"schema":{"type":"string"}}}},"500":{"description":"Error while connecting to database."}}}},"/subscription/getAllSubscriptions/{order}/{iPage}":{"get":{"security":[{"bearerAuth":[]}],"tags":["Subscription"],"parameters":[{"in":"path","name":"order","schema":{"type":"string"},"required":true,"description":"The name of the column by which the array should be sorted"},{"in":"path","name":"iPage","schema":{"type":"integer"},"required":true,"description":"The page number"}],"responses":{"200":{"description":"Returns a JSON array of 10 subscriptions, ordered by the column name chosen.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/getAllSubscriptions"}}}},"401":{"$ref":"#/components/responses/UnauthorizedError"},"403":{"$ref":"#/components/responses/mustBeAdmin"},"404":{"description":"No subscription.","content":{"text/plain":{"schema":{"type":"string"}}}},"500":{"description":"Error while connecting to database."}}}},"/subscription//getAllSubscriptionsAndPagesCount/{order}/{iPage}":{"get":{"security":[{"bearerAuth":[]}],"tags":["Subscription"],"parameters":[{"in":"path","name":"order","schema":{"type":"string"},"required":true,"description":"The name of the column by which the array should be sorted"},{"in":"path","name":"iPage","schema":{"type":"integer"},"required":true,"description":"The page number"}],"responses":{"200":{"description":"Returns a JSON object result:<br> .subscriptions: a JSON array of 10 subscriptions, ordered by the column name chosen <br> .nbPagesSubscriptions: The number of pages","content":{"application/json":{"schema":{"$ref":"#/components/schemas/getAllSubscriptionsAndPagesCount"}}}},"401":{"$ref":"#/components/responses/UnauthorizedError"},"403":{"$ref":"#/components/responses/mustBeAdmin"},"404":{"description":"No subscription.","content":{"text/plain":{"schema":{"type":"string"}}}},"500":{"description":"Error while connecting to database."}}}},"/subscription/getSearchSubscriptions/{search}/{order}/{iPage}":{"get":{"security":[{"bearerAuth":[]}],"tags":["Subscription"],"parameters":[{"in":"path","name":"search","schema":{"type":"string"},"required":true,"description":"The value of the research"},{"in":"path","name":"order","schema":{"type":"string"},"required":true,"description":"The name of the column by which the array should be sorted"},{"in":"path","name":"iPage","schema":{"type":"integer"},"required":true,"description":"The page number"}],"responses":{"200":{"description":"Returns a JSON object result:<br> .subscriptions: a JSON array of 10 subscriptions matching the search value, ordered by the column name chosen <br> .nbPagesSubscriptions: The number of pages","content":{"application/json":{"schema":{"$ref":"#/components/schemas/getAllSubscriptionsAndPagesCount"}}}},"401":{"$ref":"#/components/responses/UnauthorizedError"},"403":{"$ref":"#/components/responses/mustBeAdmin"},"404":{"description":"No subscription.","content":{"text/plain":{"schema":{"type":"string"}}}},"500":{"description":"internal server error."}}}},"/subscription/add":{"post":{"security":[{"bearerAuth":[]}],"tags":["Subscription"],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/addSubscription"}}}},"responses":{"201":{"$ref":"#/components/responses/subscriptionAdded"},"400":{"description":"the error(s) described","content":{"text/plain":{"schema":{"type":"string"}}}},"401":{"$ref":"#/components/responses/UnauthorizedError"},"403":{"$ref":"#/components/responses/mustBeAdmin"},"500":{"description":"internal server error."}}}},"/subscription/update":{"patch":{"security":[{"bearerAuth":[]}],"tags":["Subscription"],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/updateSubscription"}}}},"responses":{"200":{"description":"Returns a JSON array of 10 subscriptions, ordered by the column name chosen.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/getAllSubscriptions"}}}},"401":{"$ref":"#/components/responses/UnauthorizedError"},"403":{"$ref":"#/components/responses/mustBeAdmin"},"404":{"description":"No subscription.","content":{"text/plain":{"schema":{"type":"string"}}}},"500":{"description":"Error while connecting to database."}}}},"/subscription/delete":{"delete":{"security":[{"bearerAuth":[]}],"tags":["Subscription"],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/deleteSchema"}}}},"responses":{"200":{"description":"Returns a JSON array of 10 subscriptions, ordered by the column name chosen.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/getAllSubscriptions"}}}},"401":{"$ref":"#/components/responses/UnauthorizedError"},"403":{"$ref":"#/components/responses/mustBeAdmin"},"404":{"description":"No subscription.","content":{"text/plain":{"schema":{"type":"string"}}}},"500":{"description":"Error while connecting to database."}}}}},"components":{"schemas":{"Sponsoring":{"type":"object","properties":{"referred":{"type":"integer","description":"The referred id"},"sponsor":{"type":"integer","description":"The sponsor id"}}},"getAllSponsoring":{"type":"array","items":{"$ref":"#/components/schemas/Sponsoring"}},"getAllSponsoringAndPagesCount":{"type":"object","properties":{"sponsoring":{"$ref":"#/components/schemas/getAllSponsoring"},"nbPagesSponsoring":{"type":"integer"}}},"Subscription":{"type":"object","properties":{"id":{"type":"integer"},"label":{"type":"string"},"price":{"type":"number"},"discount":{"type":"number","minimum":0,"maximum":1},"paymentRecurrence":{"type":"string"},"vehicleType":{"type":"string"}}},"getAllSubscriptions":{"type":"array","items":{"$ref":"#/components/schemas/Subscription"}},"getAllSubscriptionsAndPagesCount":{"type":"object","properties":{"subscriptions":{"$ref":"#/components/schemas/getAllSubscriptions"},"nbPagesSubscriptions":{"type":"integer"}}},"addSponsoring":{"type":"object","properties":{"iPage":{"type":"integer"},"column":{"type":"string"},"referred":{"type":"integer"},"sponsor":{"type":"integer"}},"required":["iPage","column","referred","sponsor"]},"updateSponsoring":{"type":"object","properties":{"iPage":{"type":"integer"},"column":{"type":"string"},"referred":{"type":"integer"},"sponsor":{"type":"integer"}},"required":["iPage","column","referred"]},"addSubscription":{"type":"object","properties":{"iPage":{"type":"integer"},"column":{"type":"string"},"label":{"type":"string"},"price":{"type":"number"},"discount":{"type":"number","minimum":0,"maximum":1},"paymentRecurrence":{"type":"string"},"vehicleType":{"type":"string"}},"required":["iPage","column","label","price","discount","paymentRecurrence","vehicleType"]},"updateSubscription":{"type":"object","properties":{"iPage":{"type":"integer"},"column":{"type":"string"},"id":{"type":"integer"},"label":{"type":"string"},"price":{"type":"number"},"discount":{"type":"number","minimum":0,"maximum":1},"paymentRecurrence":{"type":"string"},"vehicleType":{"type":"string"}},"required":["iPage","column","id"]},"deleteSchema":{"type":"array","items":{"type":"integer"}}},"responses":{"sponsoringAdded":{"description":"Returns a JSON object result:<br> .id: The id<br> .sponsoring: a JSON array of 10 sponsoring matching the search value, ordered by the column name chosen <br> .nbPagesSponsoring: The number of pages","content":{"application/json":{"schema":{"type":"object","properties":{"id":{"type":"integer"},"sponsoring":{"$ref":"#/components/schemas/getAllSponsoring"},"nbPagesSponsoring":{"type":"integer"}}}}}},"subscriptionAdded":{"description":"Returns a JSON object result:<br> .id: The id<id> .subscriptions: a JSON array of 10 subscriptions matching the search value, ordered by the column name chosen <br> .nbPagesSubscriptions: The number of pages","content":{"application/json":{"schema":{"type":"object","properties":{"id":{"type":"integer"},"subscriptions":{"$ref":"#/components/schemas/getAllSubscriptions"},"nbPagesSubscriptions":{"type":"integer"}}}}}},"mustBeAdmin":{"description":"The action must be realized by an admin."},"UnauthorizedError":{"description":"JWT is missing or invalid","content":{"text/plain":{"schema":{"type":"string"}}}}},"securitySchemes":{"bearerAuth":{"type":"http","scheme":"bearer","bearerFormat":"JWT"}}},"tags":[]}